backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Paperazzi Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "pillow>=10.4.0",
    "python-multipart>=0.0.20",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.0",
    "gradium>=0.5.11",
    "pymupdf>=1.27.2.3",
    "docling>=2.91.0",
]
